Added initial Gitea Action for oms_sensemaking (linting only for now).
The Gitea Action is currently on running ruff on the code base. There
is a job for running unit tests, but it is comment out, while we
figure out how to better handle the NLP tests. If we do want to run
the tests as-is, we will need to explore options for installing
or accessing the CoreNLP as part of the unit tests.