Fix pgadmin not starting
This was due to unmet dependencies on the ess-indexer-service,
which was part of the global docker compose profile, but depended
on omsb-init which was only part of the `local` profile.  The
solution here is to add the ES related dependencies to the `local`
profile.  They're only there for storage for OMSB, so they should
only be available when it is running.